Refactor color handling
Now a custom class ColorButton is used in the toolbar. This
enables some significant code simplifications. And there is
now a list of colors. This list can be of different size. So
everything is prepared for more than 10 colors.

But the real problem is, of course, backwards compatibility
with db-bead. I have to be able to load db-bead files correctly.
And I have to be able to save jbead files with up to 10 colors
so that db-bead can reasonably load and display them. Files with
more than 10 colors will not be readable by db-bead. Such files
will have a totally rewritten file format, most probably.

+++ src/ch/jbead/BasePanel.java
@@ -51,7 +51,7 @@ public abstract class BasePanel extends JComponent implements ModelListener {
     }
 
     @Override
-    public void colorChanged(int colorIndex) {
+    public void colorChanged(byte colorIndex) {
         repaint();
     }
