+++ src/ch/jbead/ModelListener.java
@@ -18,13 +18,13 @@
 package ch.jbead;
 
 /**
- * 
+ *
  */
 public interface ModelListener {
 
     public void pointChanged(Point pt);
     public void modelChanged();
-    public void colorChanged(int colorIndex);
+    public void colorChanged(byte colorIndex);
     public void scrollChanged(int scroll);
     public void shiftChanged(int shift);
     public void zoomChanged(int gridx, int gridy);
